+++ react-components/src/components/ProductFinder/CountryFinder.jsx
@@ -1,18 +1,19 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import ReactDOM from 'react-dom'
 import ReactModal from 'react-modal'
-import { getModalIsOpen, getProductsExpertise } from '@src/reducers'
+import PropTypes from 'prop-types'
 import Services from '@src/Services'
-import MessageConfirmation from './MessageConfirmation'
-import RegionToggle from "./RegionToggle";
+import RegionToggle from './RegionToggle'
+import Confirmation from './MessageConfirmation'
 
 
 const suggested = [{'country':'France', 'region':'Europe'}, {'country': 'Spain', 'region':'Europe'}, {'country':'Italy', 'region':'Europe'}, {'country':'Jamaica',  'region':'Latin America and Caribbean'}]
 
 export function CountryFinder(props) {
   let modalContent
+  const { text } = props
   const [modalIsOpen, setIsOpen] = useState(false)
-  const [selectedCountry, setSelectedCountry] = useState(props.text)
+  const [selectedCountry, setSelectedCountry] = useState(text)
   const [countryList, setCountryList] = useState()
   const [searchStr, setSearchStr] = useState()
   const [productConfirmationRequired, setProductConfirmationRequired] = useState(false)
@@ -34,16 +35,8 @@ export function CountryFinder(props) {
     setIsOpen(true)
   }
 
-  const modalAfterOpen = () => {
-    modalContent.style.maxHeight = '700px'
-    if (!countryList) {
-      getCountries()
-    }
-  }
-
   const searchChange = (evt) => {
-    let searchString = evt.target.value
-    setSearchStr(searchString.toUpperCase())
+    setSearchStr(evt.target.value.toUpperCase())
     setExpandRegion(true)
   }
 
@@ -54,13 +47,21 @@ export function CountryFinder(props) {
   const getCountries = () => {
     Services.getCountries().then((result) => {
       // map regions
-      let regions = {}
-      for (const [index, country] of result.entries()) {
-        let region = country.region;
+      const regions = {}
+      result.map((country) => {
+        const { region } = country;
         (regions[region] = regions[region] || []).push(country)
-      }
+        return null
+      })
       setCountryList(regions)
     })  
+  }
+
+  const modalAfterOpen = () => {
+    modalContent.style.maxHeight = '700px'
+    if (!countryList) {
+      getCountries()
+    }
   }
 
   const saveCountry = (country) => {
@@ -76,14 +77,14 @@ export function CountryFinder(props) {
     })
       .then((result) => {
         closeModal()
+        window.location.reload()
       })
-      .catch((result) => {
+      .catch(() => {
         // TODO: Add error confirmation here
       })
   }
 
   const selectCountry = (evt) => {
-    let targetCountry = evt.target.getAttribute('data-country')
     saveCountry({
       name: evt.target.getAttribute('data-country'),
       id: evt.target.getAttribute('data-id'),
@@ -91,15 +92,13 @@ export function CountryFinder(props) {
     })
   }
 
-  let _regions = Object.keys(countryList || {}).map((region, index) => {
-    let countryFound = false
-    let _countries = (countryList[region] || []).map((country, index) => {
-      countryFound = true
+  let regions = Object.keys(countryList || {}).map((region, index) => {
+    let countries = (countryList[region] || []).map((country, index) => {
       if (searchStr && country.name.toUpperCase().indexOf(searchStr) != 0) return ''
       return (
         <span className="c-1-5" key={index}>
           <li>
-            <button type="button" className="link m-r-s m-b-xs" data-country={country.name} data-id={country.id} data-region={country.region}>
+            <button type="button" className="link m-r-s m-b-xs" data-country={country.name} data-id={country.id} data-region={country.region} onClick={selectCountry}>
               {country.name}
             </button>
           </li>
@@ -107,30 +106,30 @@ export function CountryFinder(props) {
       )
     })
     return (
-      !!_countries.filter((region) => region).length && (
-        <RegionToggle key={index} expandAllRegions={expandRegion} region={region} countries={_countries} />
+      !!countries.filter((countryRegion) => countryRegion).length && (
+        <RegionToggle key={index} expandAllRegions={expandRegion} region={region} countries={countries} />
       )
     )
   })
 
-  if (!_regions.filter((region) => region).length) {
-    _regions = <div className="h-xs">No results found</div>
+  if (!regions.filter((region) => region).length) {
+    regions = <div className="h-xs">No results found</div>
   }
-  const _suggested = []
-  for (const [index, value] of suggested.entries() ){
-    _suggested.push(
-      <button key={index} type="button" className="tag tag--tertiary tag--icon m-r-s" data-country={value.country} data-region={value.region}>
+  const suggestedList = suggested.map((value) => {
+    return (
+      <button key={`suggested_${value.country}`} type="button" className="tag tag--tertiary tag--icon m-r-s" data-country={value.country}  data-region={value.region} onClick={selectCountry}>
         {value.country}
       </button>
     )
-  }
-  let buttonClass = 'tag ' + (!selectedCountry ? 'tag--tertiary' : '') + ' tag--icon '
+  })
+
+  const buttonClass = `tag ${!selectedCountry ? 'tag--tertiary' : ''} tag--icon `
 
   return (
     <span>
-      <button className={buttonClass} onClick={openModal}>
+      <button type="button" className={buttonClass} onClick={openModal}>
         {selectedCountry || 'add country'}
-        <i className={'fa ' + (selectedCountry ? 'fa-edit' : 'fa-plus')}></i>
+        <i className={`fa ${(selectedCountry ? 'fa-edit' : 'fa-plus')}`}/>
       </button>
       <ReactModal
         isOpen={modalIsOpen}
@@ -138,11 +137,11 @@ export function CountryFinder(props) {
         className="modal max-modal"
         overlayClassName="modal-overlay center"
         onAfterOpen={modalAfterOpen}
-        contentRef={(_modalContent) => (modalContent = _modalContent)}
+        contentRef={(_modalContent) => {modalContent = _modalContent; return null}}
       >
         <form className="country-chooser">
-          <div className="modal-header" style={{ height: '100px' }}>
-            <button className="pull-right m-r-0 dialog-close" onClick={closeModal}></button>
+          <div style={{ height: '100px' }}>
+            <button type="button" className="pull-right m-r-0 dialog-close" aria-label="Close" onClick={closeModal}/>
             <h2 className="h-m p-v-xs">Choose a target market</h2>
           </div>
           <div className="scroll-area" style={{ marginTop: '100px' }}>
@@ -151,22 +150,22 @@ export function CountryFinder(props) {
               <p className="m-v-xs">
                 These are based on the size of the market for your product, export distance, tariffs and costs.
               </p>
-              <ul className="m-v-xs" onClick={selectCountry}>
-                {_suggested}
+              <ul className="m-v-xs">
+                {suggestedList}
               </ul>
-              <hr className="bg-black-70"></hr>
+              <hr className="bg-black-70"/>
 
               <h3 className="h-s p-t-0">Compare markets</h3>
               <div className="grid">
                 <div className="c-full">
                   <p className="m-v-xs">Compare stats for over 180 markets to find the best place to export.</p>
-                  <a href="#" className="button button--secondary">
+                  <a href="/" className="button button--secondary">
                     Compare markets
                   </a>
                 </div>
               </div>
 
-              <hr className="bg-black-70"></hr>
+              <hr className="bg-black-70"/>
               <h3 className="h-s p-t-0">List of markets</h3>
               <p className="m-v-xs">
                 If you have an idea of where you want to export, choose from the list below. You can change this at any
@@ -192,7 +191,7 @@ export function CountryFinder(props) {
                   <button type="button" key="region-expand" className="region-expand link" onClick={toggleRegion}>{expandRegion ? 'Collapse all' : 'Expand all' }</button>
                     <hr/>
                   <ul className="country-list" onClick={selectCountry}>
-                    {_regions}
+                    {regions}
                   </ul>
                 </div>
               </div>
@@ -200,11 +199,11 @@ export function CountryFinder(props) {
           </div>
         </form>
       </ReactModal>
-      <MessageConfirmation
+      <Confirmation
         buttonClass={buttonClass}
         productConfirmation={productConfirmationRequired}
         handleButtonClick={closeConfirmation}
-        messsageTitle="Changing target market?"
+        messageTitle="Changing target market?"
         messageBody="if you've created an export plan, make sure you update it to reflect your new market. you can change target market at any time."
         messageButtonText="Got it"
       />
@@ -212,10 +211,17 @@ export function CountryFinder(props) {
   )
 }
 
-export default function({ ...params }) {
+CountryFinder.propTypes = {
+  text: PropTypes.string
+}
+CountryFinder.defaultProps = {
+  text: null
+}
+
+export default function createCountryFinder({ ...params }) {
   const mainElement = document.createElement('span')
   document.body.appendChild(mainElement)
   ReactModal.setAppElement(mainElement)
-  let text = params.element.getAttribute('data-text')
-  ReactDOM.render(<CountryFinder text={text}></CountryFinder>, params.element)
+  const text = params.element.getAttribute('data-text')
+  ReactDOM.render(<CountryFinder text={text}/>, params.element)
 }