+++ react-components/src/components/ProductFinder/Interaction.jsx
@@ -29,7 +29,7 @@ function RadioButtons(props) {
     const checked = option.id === (selection && selection.id)
     const optionName = (<span>{(option.isPart ? <span className="text-black-50">Part of </span> : '')}{capitalize(option.name, !option.isPart)}</span>)
     return (
-      <label key={option.id} htmlFor={option.id} className="multiple-choice p-f-m m-b-xxs">
+      <label key={option.id} htmlFor={option.id} className="multiple-choice p-f-m m-v-xxs">
         <input
           type="radio"
           className="radio"
@@ -47,7 +47,7 @@ function RadioButtons(props) {
       </label>
     )
   })
-  return <div className="m-b-xs" style={{overflow:'hidden'}} onChange={changeVal}>{buttons}</div>
+  return <div className="m-b-xs" onChange={changeVal}>{buttons}</div>
 }
 
 RadioButtons.propTypes = {
@@ -61,8 +61,6 @@ RadioButtons.propTypes = {
 
 
 export default function Interaction(props) {
-  const buttonEnabled = true
-
   const { txId, attribute, isItemChoice, processResponse } = props
 
   const [value, setValue] = useState()
@@ -74,25 +72,24 @@ export default function Interaction(props) {
       processResponse(
         Services.lookupProductRefine({
           txId,
-          attributeId: attribute.id,
-          valueId: value.id,
-          valueString: value.name
+          interactionId: attribute.id,
+          values:[{first: value.id, second: value.name}]
         })
       )
     }
   }
 
-  const valueChange = (_value) => {
-    setValue(_value)
+  const valueChange = (newValue) => {
+    setValue(newValue)
   }
 
   return (
     <div className="interaction grid m-v-xs" key={attribute.id}>
         <div className="c-fullwidth">
-          <span className="interaction-name h-xs p-t-0">{capitalize(attribute.label)}</span>
-          <p className="m-v-xxs">Select the best match for your product.</p>
+          <span className="interaction-name h-s p-t-0">{capitalize(attribute.label)}</span>
+          <p className="m-v-xs">Select the best match for your product.</p>
           <RadioButtons attribute={attribute} valueChange={valueChange}/>
-          <button type="button" className="button button--secondary" disabled={!buttonEnabled} onClick={clickNext}>Next</button>
+          <button type="button" className="button button--secondary m-t-xxs" disabled={!value || !Object.keys(value).length} onClick={clickNext} style={{float:'left',clear:'both'}}>Next</button>
         </div>
     </div>
   )