+++ CHANGELOG.md
@@ -59,6 +59,9 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-261 - Additional lesson video/image block
+- GP2-707 - Capability design review
+- GP2-684 - Product-search value interaction
 - GP2-669 - CountryFinder added collapsibility for region section
 - GP2-703 - Add new DSM Cards
 - GP2-686 - Product-search End result details
@@ -261,3 +264,5 @@
 - GP2-655 - local development support/skeleton for Wagtail Transfer work
 - GP2-667 - appropriately scoped configuration for Wagtail Transfer when deployed
 - GP2-656 - Make transcript field mandatory for GreatMedia
+- GP2-709 - BE Country add required country signal
+- GP2-744 and GP2-744 - tighten up validation for CaseStudy.body StreamField